feat: server owner messaging
This adds the ability to send messages to the server owner.
This is useful for reminding owners to complete their bot setup, for issues, and even notifications.

For now we use the feature to just send a server setup message that pings the user with details about the created channels, and how to get started with the FCM credentials.